pyproject.toml
[project]
name = "topic-modeling-streamlit"
version = "0.1.2"
description = ""
authors = [
    { name = "Bor Hodošček", email = "dev@bor.space" }
]
license = "MIT"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "accelerate>=1.8.1",
    "bertopic>=0.17.3",
    "bitsandbytes>=0.46.1",
    "boto3>=1.39.13",
    "datamapplot>=0.6.3",
    "einops>=0.8.1",
    "en-core-web-sm",
    "en-core-web-trf",
    "faiss-gpu-cu12>=1.11.0",
    "filelock>=3.18.0",
    "fugashi-plus>=1.4.0.post1",
    "gensim==4.3.3",
    "ipython>=9.4.0",
    "ja-core-news-sm",
    "jaconv>=0.4.0",
    "litellm>=1.74.7",
    "matplotlib>=3.10.3",
    "numba>=0.61.2",
    "numpy>=1.26.4",
    # "pandas>=2.3.1",
    "polars>=1.31.0",
    # "pyarrow>=20.0.0",
    "pydantic>=2.11.7",
    "pyldavis>=3.4.1",
    "pysbd>=0.3.4",
    "regex>=2024.11.6",
    "rhoknp>=1.7.1",
    "safetensors>=0.5.3",
    "scipy>1.7",
    "sentencepiece>=0.2.0",
    "spacy>=3.8.7",
    "stqdm>=0.0.5",
    "streamlit>=1.46.1",
    "sudachidict-full>=20250515",
    "sudachipy>=0.6.10",
    "tenacity>=9.1.2",
    "tiktoken>=0.9.0",
    "torch<2.8.0", # Need new flash-attn version to bump
    "torchdr>=0.3",
    "umap-learn>=0.5.9.post2",
    "unidic-lite>=1.0.8",
    "xxhash>=3.5.0",
]

[tool.uv]
no-build-isolation-package = ["flash-attn"]

environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
en-core-web-trf = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl" }
ja-core-news-sm = { url = "https://github.com/explosion/spacy-models/releases/download/ja_core_news_sm-3.8.0/ja_core_news_sm-3.8.0-py3-none-any.whl" }

cudf-cu12 = { index = "cuml" }
cuml-cu12 = { index = "cuml" }
rapids-dask-dependency = { index = "cuml" }
dask = { index = "cuml" }

[[tool.uv.index]]
name = "cuml"
url = "https://pypi.nvidia.com"
explicit = true

[[tool.uv.dependency-metadata]]
name = "flash-attn"
version = "2.8.2"
requires-dist = ["torch", "einops"]

[project.optional-dependencies]
build = ["torch", "setuptools", "packaging"]
compile = ["flash-attn"]
cuda = [
    "cupy-cuda12x>=13.5.1",
    "flash-attn>=2.8.1",
]

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "pandas-stubs>=2.3.0.250703",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "types-tqdm>=4.67.0.20250516",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/topic_modeling_streamlit"]

[tool.pytest.ini_options]
minversion = "6.0"
# addopts = "-ra -q --doctest-modules --cov-report term-missing:skip-covered --cov=topic_modeling_streamlit"
testpaths = [
    "tests",
    "src",
]
doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL"

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src", "tests"]
